+++ tea_django/models/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["VanillaModel", "UUIDBaseModel"]
+
+from tea_django.models.vanilla import VanillaModel
+from tea_django.models.uuid_base import UUIDBaseModel