+++ tea_django/models/mixins/__init__.py
@@ -1,0 +1,12 @@
+__all__ = [
+    "ColoredMixin",
+    "UniqueSlugMixin",
+    "NonUniqueSlugMixin",
+    "TimerMixin",
+    "TimestampedMixin",
+]
+
+from tea_django.models.mixins.colored import ColoredMixin
+from tea_django.models.mixins.slug import UniqueSlugMixin, NonUniqueSlugMixin
+from tea_django.models.mixins.timer import TimerMixin
+from tea_django.models.mixins.timestamped import TimestampedMixin