+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # CHANGELOG
 
 
+## v0.0.5 (August 24, 2020)
+
+- Fix secret key saving into config file.
+- Rename project to `tea-django`.
+
+
+---
+
+
 ## v0.0.4 (August 23, 2020)
 
 - Fixes.
@@ -10,7 +19,7 @@
 
 
 ## v0.0.3 (August 23, 2020)
-- Move stuff to `console_tea` and `tea` projects. 
+- Move stuff to `tea-console` and `tea` projects. 
 
 
 ---